Fixing github commits. Removing bag files and pushing.
