+++ cart_endpoints/scripts/passenger/passenger.py
@@ -1,10 +1,13 @@
 #!/usr/bin/env python
 import rospy
+import numpy as np
+import pandas as pd
 from zed_interfaces.msg import ObjectsStamped
 from geometry_msgs.msg import TwistStamped, Vector3, PointStamped, TransformStamped
 from sensor_msgs.msg import Image
-from std_msgs.msg import Bool, Int8
+from std_msgs.msg import Bool, Int8, String
 from navigation_msgs.msg import VehicleState, Stop, VelAngle
+from passengerclassifier import PassengerRFClassifier, PassengerData
 
 '''
 This ROS node keeps track of passenger pose data and determines whether passengers
@@ -15,12 +18,15 @@ Version: 04/2023
 '''
 PASSENGER_EDGE_TOP_X_THRESHOLD = 0.5
 DRIVER_EDGE_TOP_X_THRESHOLD = -0.7
-COUNT_THRESHOLD = 5 # in frames
+COUNT_THRESHOLD = 0 # in frames
+UNSAFE_COUNT_THRESHOLD = 40
+STOP_COUNT_THRESHOLD = 5
 DEPTH_THRESHOLD = 1.3
+
 
 class Passenger(object):
     def __init__(self):
-        rospy.init_node('Passenger')
+        rospy.init_node('Passenger_ML')
         
         # Topic that object detection data from passenger camera is coming from
         self.objects_in = rospy.get_param("objects_in", "/passenger_cam/passenger/obj_det/objects")
@@ -28,10 +34,10 @@ class Passenger(object):
         # Publishers
         self.out_of_bounds_pub = rospy.Publisher('/passenger/out_of_bounds', Bool, queue_size=10)
         self.occupants_pub = rospy.Publisher('/passenger/occupants', Int8, queue_size=10)
-        self.stop_pub = rospy.Publisher('/passenger/emergency_stop', Bool, queue_size=10)
+        self.stop_pub = rospy.Publisher('/passenger/emergency_stop', String, queue_size=10)
 
         # subscribers:
-        self.object_sub = rospy.Subscriber(self.objects_in, ObjectsStamped, callback=self.received_persons, queue_size=10)
+        self.object_sub = rospy.Subscriber(self.objects_in, ObjectsStamped, callback=self.received_persons, queue_size=1)
         self.nav_sub = rospy.Subscriber('/nav_cmd', VelAngle, self.nav_cb, queue_size=10)
 
         #Private Variables
@@ -40,6 +46,7 @@ class Passenger(object):
         self.startingOccupants = 1
         self.stop_counter = 0
         self.stopped = True
+        self.prfc = PassengerRFClassifier.load_model("./saved_models/passengerRF_model_113")
 
         rospy.loginfo("Started pose tracking node! (S23)")
 
@@ -69,23 +76,29 @@ class Passenger(object):
             passenger_edge = person.skeleton_3d.keypoints[2].kp[1]  # passenger's top right side, y point
             person_depth = person_corners[1].kp[0]                  # occupant's furthest back point, z position
 
+            
+
             # Ignore passengers beyond a certain depth
             if person_depth > DEPTH_THRESHOLD:
                 continue
 
+            if passenger_edge < DRIVER_EDGE_TOP_X_THRESHOLD or driver_edge > PASSENGER_EDGE_TOP_X_THRESHOLD:
+                continue
+            
+            occupant_count += 1
+
+            passenger_keypoints = []
+            for kp in person.skeleton_3d.keypoints:
+                passenger_keypoints.append(kp.kp)
+
+            passenger_keypoints = np.reshape(np.array([passenger_keypoints]), (-1, 18*3))
+            prediction = self.prfc.predict(passenger_keypoints)
+
             # Classify passengers based on position
-            if (driver_edge < DRIVER_EDGE_TOP_X_THRESHOLD  \
-                    and passenger_edge > DRIVER_EDGE_TOP_X_THRESHOLD) \
-                    or (passenger_edge > PASSENGER_EDGE_TOP_X_THRESHOLD \
-                    and driver_edge < PASSENGER_EDGE_TOP_X_THRESHOLD):
+            if prediction == 0:
                 # Passenger is crossing threshold, signifying an unsafe occupant
                 unsafe_person = True
-                occupant_count += 1
-                self.has_seen_passanger = True
 
-            elif (driver_edge > DRIVER_EDGE_TOP_X_THRESHOLD and passenger_edge < PASSENGER_EDGE_TOP_X_THRESHOLD): 
-                # Passenger is within the threshold, signifying a safe occupant
-                occupant_count += 1
         self.occupants = occupant_count
                 
         # Iterate stop_count
@@ -95,18 +108,23 @@ class Passenger(object):
             self.stop_counter = 0
 
         # Iterate out_count
-        if not unsafe_person: #and self.has_seen_passanger:
-            self.out_counter = 0
-        else:
+        if unsafe_person and not self.stopped:
             self.out_counter += 1
+        else:
+            self.out_counter = 0
+        
+        # Check if a pose is unsafe for too long.
+        if self.out_counter > UNSAFE_COUNT_THRESHOLD:
+            self.stop_pub.publish("unsafe-pose-stop")
+        
+        #print("UNSAFE COUNT: ", self.out_counter, "/", UNSAFE_COUNT_THRESHOLD)
 
         # Publish passenger information
-        if self.stop_counter > COUNT_THRESHOLD and not self.stopped:        # Publish Emergency Stop
-            self.stop_pub.publish(True)
+        if self.stop_counter > STOP_COUNT_THRESHOLD and not self.stopped:        # Publish Emergency Stop
+            self.stop_pub.publish("passenger-exit-stop")
             self.stop_counter = 0 
         self.occupants_pub.publish(self.occupants)                          # Publish Occupant Count
         self.out_of_bounds_pub.publish(self.out_counter > COUNT_THRESHOLD)  # Publish Out of Bounds
-
     ''' 
     Call back method for detecting whether cart is stopped or not 
     ____________
